order_service_az/requirements.txt
flask
flask-restful
azure-cosmos
uuid